sys/backend/api/requirements.txt
django>=6.0,<7.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.9,<5.0
django-redis>=6.0,<7.0
psycopg2-binary>=2.9,<3.0
celery>=5.6,<6.0
redis>=7.0,<8.0
whitenoise>=6.8,<7.0

# Testing
pytest>=8.3,<9.0
pytest-django>=4.10,<5.0
pytest-cov>=6.0,<7.0
factory-boy>=3.3,<4.0
Faker>=24.0,<25.0
